refactor: move stage6 relation state out of jolt-claims
