+++ crates/jolt-claims/src/protocols/jolt/formulas/bytecode.rs
@@ -2,7 +2,6 @@ use jolt_field::{Field, RingCore};
 use jolt_lookup_tables::{InstructionLookupTable, LookupTableKind, XLEN};
 use jolt_poly::{
     boolean_index_msb, eq_index_msb, EqPolynomial, IdentityPolynomial, MultilinearEvaluation,
-    Polynomial,
 };
 use jolt_riscv::{
     instructions::Noop, CircuitFlags, Flags, InstructionFlags, InterleavedBitsMarker,
@@ -454,137 +453,6 @@ pub struct BytecodeReadRafStageValueInputs<'a, F> {
 pub struct BytecodeReadRafAddressPhaseIdentityCoefficients<F> {
     pub stage1: F,
     pub stage3: F,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct BytecodeReadRafAddressPhaseState<F: Field> {
-    log_k: usize,
-    stage_f: Vec<Polynomial<F>>,
-    stage_val: Vec<Polynomial<F>>,
-    entry_trace: Polynomial<F>,
-    entry_expected: Polynomial<F>,
-    gamma_powers: Vec<F>,
-}
-
-impl<F: Field> BytecodeReadRafAddressPhaseState<F> {
-    pub fn new(
-        log_k: usize,
-        stage_f: Vec<Polynomial<F>>,
-        stage_val: Vec<Polynomial<F>>,
-        entry_trace: Polynomial<F>,
-        entry_expected: Polynomial<F>,
-        gamma_powers: Vec<F>,
-    ) -> Self {
-        Self {
-            log_k,
-            stage_f,
-            stage_val,
-            entry_trace,
-            entry_expected,
-            gamma_powers,
-        }
-    }
-
-    pub const fn log_k(&self) -> usize {
-        self.log_k
-    }
-
-    pub fn bind(&mut self, challenge: F) {
-        for polynomial in &mut self.stage_f {
-            polynomial.bind(challenge);
-        }
-        for polynomial in &mut self.stage_val {
-            polynomial.bind(challenge);
-        }
-        self.entry_trace.bind(challenge);
-        self.entry_expected.bind(challenge);
-    }
-
-    pub fn round_rows(&self) -> usize {
-        self.entry_trace.len() / 2
-    }
-
-    pub fn round_eval(&self, index: usize, point: F) -> F {
-        let mut output = F::zero();
-        for stage in 0..self.stage_f.len() {
-            output += self.gamma_powers[stage]
-                * self.stage_f[stage].sumcheck_round_eval(index, point)
-                * self.stage_val[stage].sumcheck_round_eval(index, point);
-        }
-        output
-            + self.gamma_powers[7]
-                * self.entry_trace.sumcheck_round_eval(index, point)
-                * self.entry_expected.sumcheck_round_eval(index, point)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct BytecodeReadRafRelationState<F: Field> {
-    public_coeff: Polynomial<F>,
-    bytecode_ra: Vec<Polynomial<F>>,
-    address_phase: Option<BytecodeReadRafAddressPhaseState<F>>,
-}
-
-impl<F: Field> BytecodeReadRafRelationState<F> {
-    pub fn new(
-        public_coeff: Polynomial<F>,
-        bytecode_ra: Vec<Polynomial<F>>,
-        address_phase: Option<BytecodeReadRafAddressPhaseState<F>>,
-    ) -> Self {
-        Self {
-            public_coeff,
-            bytecode_ra,
-            address_phase,
-        }
-    }
-
-    pub fn round_degree(&self, local_round: usize, fallback: usize) -> usize {
-        if self.is_address_phase_round(local_round) {
-            2
-        } else {
-            fallback
-        }
-    }
-
-    pub fn bind(&mut self, local_round: usize, challenge: F) {
-        self.public_coeff.bind(challenge);
-        for polynomial in &mut self.bytecode_ra {
-            polynomial.bind(challenge);
-        }
-        if let Some(address_phase) = &mut self.address_phase {
-            if local_round < address_phase.log_k() {
-                address_phase.bind(challenge);
-            }
-        }
-    }
-
-    pub fn round_rows(&self, local_round: usize) -> usize {
-        if let Some(address_phase) = &self.address_phase {
-            if local_round < address_phase.log_k() {
-                return address_phase.round_rows();
-            }
-        }
-        self.public_coeff.len() / 2
-    }
-
-    pub fn round_eval(&self, local_round: usize, index: usize, point: F) -> F {
-        if let Some(address_phase) = &self.address_phase {
-            if local_round < address_phase.log_k() {
-                return address_phase.round_eval(index, point);
-            }
-        }
-        let public_coeff = self.public_coeff.sumcheck_round_eval(index, point);
-        let ra_product = self.bytecode_ra.iter().fold(F::one(), |acc, polynomial| {
-            acc * polynomial.sumcheck_round_eval(index, point)
-        });
-        public_coeff * ra_product
-    }
-
-    fn is_address_phase_round(&self, local_round: usize) -> bool {
-        self.address_phase
-            .as_ref()
-            .is_some_and(|address_phase| local_round < address_phase.log_k())
-    }
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -1358,107 +1226,6 @@ mod tests {
                 Fr::from_u64(11),
             ]
         );
-    }
-
-    #[test]
-    fn read_raf_address_phase_state_evaluates_round_formula() {
-        let point = Fr::from_u64(3);
-        let stage_f = (0..5)
-            .map(|stage| {
-                Polynomial::new(vec![
-                    Fr::from_u64(stage as u64 + 2),
-                    Fr::from_u64(stage as u64 + 7),
-                ])
-            })
-            .collect::<Vec<_>>();
-        let stage_val = (0..5)
-            .map(|stage| {
-                Polynomial::new(vec![
-                    Fr::from_u64(stage as u64 + 11),
-                    Fr::from_u64(stage as u64 + 19),
-                ])
-            })
-            .collect::<Vec<_>>();
-        let entry_trace = Polynomial::new(vec![Fr::from_u64(29), Fr::from_u64(31)]);
-        let entry_expected = Polynomial::new(vec![Fr::from_u64(37), Fr::from_u64(41)]);
-        let gamma_powers = (0..8)
-            .map(|value| Fr::from_u64(value as u64 + 43))
-            .collect::<Vec<_>>();
-
-        let state = BytecodeReadRafAddressPhaseState::new(
-            1,
-            stage_f.clone(),
-            stage_val.clone(),
-            entry_trace.clone(),
-            entry_expected.clone(),
-            gamma_powers.clone(),
-        );
-        let expected = stage_f
-            .iter()
-            .zip(stage_val.iter())
-            .zip(gamma_powers.iter())
-            .map(|((stage_f, stage_val), gamma)| {
-                *gamma
-                    * stage_f.sumcheck_round_eval(0, point)
-                    * stage_val.sumcheck_round_eval(0, point)
-            })
-            .sum::<Fr>()
-            + gamma_powers[7]
-                * entry_trace.sumcheck_round_eval(0, point)
-                * entry_expected.sumcheck_round_eval(0, point);
-
-        assert_eq!(state.log_k(), 1);
-        assert_eq!(state.round_rows(), 1);
-        assert_eq!(state.round_eval(0, point), expected);
-    }
-
-    #[test]
-    fn read_raf_relation_state_evaluates_public_coeff_times_ra_product() {
-        let point = Fr::from_u64(5);
-        let public_coeff = Polynomial::new(vec![Fr::from_u64(2), Fr::from_u64(7)]);
-        let bytecode_ra = vec![
-            Polynomial::new(vec![Fr::from_u64(11), Fr::from_u64(13)]),
-            Polynomial::new(vec![Fr::from_u64(17), Fr::from_u64(19)]),
-        ];
-        let expected = public_coeff.sumcheck_round_eval(0, point)
-            * bytecode_ra[0].sumcheck_round_eval(0, point)
-            * bytecode_ra[1].sumcheck_round_eval(0, point);
-
-        let state = BytecodeReadRafRelationState::new(public_coeff, bytecode_ra, None);
-
-        assert_eq!(state.round_degree(0, 3), 3);
-        assert_eq!(state.round_rows(0), 1);
-        assert_eq!(state.round_eval(0, 0, point), expected);
-    }
-
-    #[test]
-    fn read_raf_relation_state_uses_address_phase_for_initial_address_rounds() {
-        let point = Fr::from_u64(5);
-        let public_coeff = Polynomial::new(vec![Fr::from_u64(2), Fr::from_u64(7)]);
-        let bytecode_ra = vec![Polynomial::new(vec![Fr::from_u64(11), Fr::from_u64(13)])];
-        let address_phase = BytecodeReadRafAddressPhaseState::new(
-            1,
-            (0..5)
-                .map(|_| Polynomial::new(vec![Fr::from_u64(3), Fr::from_u64(5)]))
-                .collect(),
-            (0..5)
-                .map(|_| Polynomial::new(vec![Fr::from_u64(7), Fr::from_u64(11)]))
-                .collect(),
-            Polynomial::new(vec![Fr::from_u64(13), Fr::from_u64(17)]),
-            Polynomial::new(vec![Fr::from_u64(19), Fr::from_u64(23)]),
-            (0..8)
-                .map(|value| Fr::from_u64(value as u64 + 29))
-                .collect(),
-        );
-        let address_phase_expected = address_phase.round_eval(0, point);
-
-        let state =
-            BytecodeReadRafRelationState::new(public_coeff, bytecode_ra, Some(address_phase));
-
-        assert_eq!(state.round_degree(0, 4), 2);
-        assert_eq!(state.round_rows(0), 1);
-        assert_eq!(state.round_eval(0, 0, point), address_phase_expected);
-        assert_eq!(state.round_degree(1, 4), 4);
     }
 
     #[test]
