+++ crates/jolt-claims/src/protocols/jolt/formulas/claim_reductions/advice.rs
@@ -406,75 +406,6 @@ pub fn advice_cycle_phase_dummy_scale<F: Field>(dummy_rounds: usize) -> F {
     (0..dummy_rounds).fold(F::one(), |scale, _| scale * two_inv)
 }
 
-/// Stateful prover object for the cycle-phase advice sumcheck.
-///
-/// Tracks the sorted `(advice, eq)` polynomial pair and accumulates the `1/2`
-/// scale for each dummy (inactive) round, matching `advice_cycle_phase_dummy_scale`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct AdviceCyclePhaseRelationState<F: Field> {
-    advice: Polynomial<F>,
-    eq: Polynomial<F>,
-    col_rounds: Range<usize>,
-    row_rounds: Range<usize>,
-    scale: F,
-}
-
-impl<F: Field> AdviceCyclePhaseRelationState<F> {
-    pub fn new(
-        advice: Polynomial<F>,
-        eq: Polynomial<F>,
-        col_rounds: Range<usize>,
-        row_rounds: Range<usize>,
-    ) -> Self {
-        Self {
-            advice,
-            eq,
-            col_rounds,
-            row_rounds,
-            scale: F::one(),
-        }
-    }
-
-    pub fn bind(&mut self, local_round: usize, challenge: F) {
-        if self.is_active_round(local_round) {
-            self.advice
-                .bind_with_order(challenge, BindingOrder::LowToHigh);
-            self.eq.bind_with_order(challenge, BindingOrder::LowToHigh);
-        } else {
-            self.scale *= F::from_u64(2).inv_or_zero();
-        }
-    }
-
-    pub fn round_rows(&self, local_round: usize) -> usize {
-        if self.is_active_round(local_round) {
-            self.advice.len() / 2
-        } else {
-            self.advice.len()
-        }
-    }
-
-    pub fn round_eval(&self, local_round: usize, index: usize, point: F) -> F {
-        if self.is_active_round(local_round) {
-            self.scale
-                * self
-                    .advice
-                    .sumcheck_round_eval_with_order(index, point, BindingOrder::LowToHigh)
-                * self
-                    .eq
-                    .sumcheck_round_eval_with_order(index, point, BindingOrder::LowToHigh)
-        } else {
-            self.scale
-                * self.advice.evals()[index]
-                * self.eq.evals()[index]
-                * F::from_u64(2).inv_or_zero()
-        }
-    }
-
-    fn is_active_round(&self, local_round: usize) -> bool {
-        self.col_rounds.contains(&local_round) || self.row_rounds.contains(&local_round)
-    }
-}
-
 /// Compute the column-side and row-side active round ranges for the advice
 /// cycle-phase sumcheck, given the main/advice matrix shapes and trace order.
 fn cycle_phase_round_schedule(
