+++ crates/jolt-claims/src/protocols/jolt/formulas/instruction.rs
@@ -426,57 +426,6 @@ where
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct InstructionRaVirtualizationRelationState<F: Field> {
-    eq_cycle: Polynomial<F>,
-    gamma_powers: Vec<F>,
-    committed_instruction_ra_by_virtual: Vec<Vec<Polynomial<F>>>,
-}
-
-impl<F: Field> InstructionRaVirtualizationRelationState<F> {
-    pub fn new(
-        eq_cycle: Polynomial<F>,
-        gamma_powers: Vec<F>,
-        committed_instruction_ra_by_virtual: Vec<Vec<Polynomial<F>>>,
-    ) -> Self {
-        Self {
-            eq_cycle,
-            gamma_powers,
-            committed_instruction_ra_by_virtual,
-        }
-    }
-
-    pub fn bind(&mut self, challenge: F) {
-        self.eq_cycle.bind(challenge);
-        for group in &mut self.committed_instruction_ra_by_virtual {
-            for polynomial in group {
-                polynomial.bind(challenge);
-            }
-        }
-    }
-
-    pub fn round_rows(&self) -> usize {
-        self.eq_cycle.len() / 2
-    }
-
-    pub fn round_eval(&self, index: usize, point: F) -> F {
-        let eq = self.eq_cycle.sumcheck_round_eval(index, point);
-        let mut output = F::zero();
-        for (gamma, group) in self
-            .gamma_powers
-            .iter()
-            .copied()
-            .zip(&self.committed_instruction_ra_by_virtual)
-        {
-            let product = group.iter().fold(F::one(), |acc, polynomial| {
-                acc * polynomial.sumcheck_round_eval(index, point)
-            });
-            output += gamma * product;
-        }
-        eq * output
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
 pub struct InstructionRaVirtualizationOutputOpenings {
     pub committed_instruction_ra_by_virtual: Vec<Vec<JoltOpeningId>>,
 }
@@ -1411,92 +1360,6 @@ mod tests {
         assert_eq!(
             ra_virtualization_eq_cycle_polynomial(&read_raf_cycle).evals(),
             EqPolynomial::<Fr>::evals(&eq_point, None)
-        );
-    }
-
-    #[test]
-    fn ra_virtualization_relation_state_evaluates_eq_weighted_virtual_products() {
-        let point = Fr::from_u64(7);
-        let eq_cycle = Polynomial::new(vec![Fr::from_u64(2), Fr::from_u64(3)]);
-        let gamma_powers = vec![Fr::from_u64(1), Fr::from_u64(11), Fr::from_u64(121)];
-        let groups = vec![
-            vec![
-                Polynomial::new(vec![Fr::from_u64(5), Fr::from_u64(13)]),
-                Polynomial::new(vec![Fr::from_u64(17), Fr::from_u64(19)]),
-            ],
-            vec![
-                Polynomial::new(vec![Fr::from_u64(23), Fr::from_u64(29)]),
-                Polynomial::new(vec![Fr::from_u64(31), Fr::from_u64(37)]),
-            ],
-            vec![
-                Polynomial::new(vec![Fr::from_u64(41), Fr::from_u64(43)]),
-                Polynomial::new(vec![Fr::from_u64(47), Fr::from_u64(53)]),
-            ],
-        ];
-        let expected = eq_cycle.sumcheck_round_eval(0, point)
-            * gamma_powers
-                .iter()
-                .copied()
-                .zip(&groups)
-                .map(|(gamma, group)| {
-                    gamma
-                        * group.iter().fold(Fr::from_u64(1), |acc, polynomial| {
-                            acc * polynomial.sumcheck_round_eval(0, point)
-                        })
-                })
-                .sum::<Fr>();
-
-        let state = InstructionRaVirtualizationRelationState::new(eq_cycle, gamma_powers, groups);
-
-        assert_eq!(state.round_rows(), 1);
-        assert_eq!(state.round_eval(0, point), expected);
-    }
-
-    #[test]
-    fn ra_virtualization_relation_state_binds_all_polynomials() {
-        let challenge = Fr::from_u64(19);
-        let mut eq_cycle =
-            Polynomial::new((0..4).map(|value| Fr::from_u64(value as u64 + 2)).collect());
-        let gamma_powers = vec![Fr::from_u64(1), Fr::from_u64(11)];
-        let mut groups = vec![
-            vec![
-                Polynomial::new((0..4).map(|value| Fr::from_u64(value as u64 + 7)).collect()),
-                Polynomial::new(
-                    (0..4)
-                        .map(|value| Fr::from_u64(value as u64 + 13))
-                        .collect(),
-                ),
-            ],
-            vec![
-                Polynomial::new(
-                    (0..4)
-                        .map(|value| Fr::from_u64(value as u64 + 17))
-                        .collect(),
-                ),
-                Polynomial::new(
-                    (0..4)
-                        .map(|value| Fr::from_u64(value as u64 + 23))
-                        .collect(),
-                ),
-            ],
-        ];
-        let mut state = InstructionRaVirtualizationRelationState::new(
-            eq_cycle.clone(),
-            gamma_powers.clone(),
-            groups.clone(),
-        );
-
-        state.bind(challenge);
-        eq_cycle.bind(challenge);
-        for group in &mut groups {
-            for polynomial in group {
-                polynomial.bind(challenge);
-            }
-        }
-
-        assert_eq!(
-            state,
-            InstructionRaVirtualizationRelationState::new(eq_cycle, gamma_powers, groups)
         );
     }
 }
