+++ crates/jolt-prover/src/stages/stage6/mod.rs
@@ -2,6 +2,7 @@ mod batch;
 mod io;
 pub(crate) mod prepare;
 pub mod prove;
+mod relation_state;
 mod verifier_output;
 
 #[cfg(feature = "zk")]
